Refactored rd_collect metrics framework.
Single entry point for data collection rd_collect.sh.
Can now run against both libjpeg and mozjpeg.
Added support for x264 and x265.

+++ tools/rd_collect_libvpx.sh
@@ -1,34 +1,27 @@
 #!/bin/bash
 set -e
 
-PLANE=$1
-CODEC=$2
-VPXENC=$3
-VPXDEC=$4
-DUMP_PSNR=$5
-DUMP_PSNRHVS=$6
-DUMP_SSIM=$7
-DUMP_FASTSSIM=$8
-FILE=$9
+if [ -z $RD_COLLECT_SUB ]; then
+  echo "Please use: $(dirname $0)/rd_collect.sh <vp8|vp9> *.y4m"
+  exit 1
+fi
+
+FILE=$1
 
 BASENAME=$(basename $FILE)
 rm $BASENAME.out 2> /dev/null || true
 echo $BASENAME
+
 WIDTH=$(head -1 $FILE | cut -d\  -f 2 | tr -d 'W')
 HEIGHT=$(head -1 $FILE | cut -d\  -f 3 | tr -d 'H')
+PIXELS=$(($WIDTH*$HEIGHT))
 
-if [ $CODEC == vp8 ]; then
-  RANGE=$(seq 50 100 5000)
-  QUALITY=--target-bitrate
-else
-  RANGE=$(seq 3 63)
-  QUALITY="--end-usage=q --cq-level"
-fi
+RANGE=$(seq 1 63)
+QSTR="-y --min-q=\$x --max-q=\$x"
 
 for x in $RANGE; do
-  $VPXENC --codec=$CODEC --good --cpu-used=0 $QUALITY=$x -o $BASENAME.vpx $FILE 2> $BASENAME-$x-enc.out
+  $VPXENC --codec=$CODEC --good --cpu-used=0 $(echo $QSTR | sed 's/\$x/'$x'/g') -o $BASENAME.vpx $FILE 2> $BASENAME-$x-enc.out
   $VPXDEC --codec=$CODEC -o $BASENAME.y4m $BASENAME.vpx
-  PIXELS=$(($WIDTH*$HEIGHT))
   SIZE=$(wc -c $BASENAME.vpx | awk '{ print $1 }')
   PSNR=$($DUMP_PSNR $FILE $BASENAME.y4m 2> /dev/null | grep Total | tr -s ' ' | cut -d\  -f $((4+$PLANE*2)))
   PSNRHVS=$($DUMP_PSNRHVS $FILE $BASENAME.y4m 2> /dev/null | grep Total | tr -s ' ' | cut -d\  -f $((4+$PLANE*2)))
