+++ tools/rd_collect_x264.sh
@@ -1,36 +1,33 @@
 #!/bin/bash
 set -e
 
-PLANE=$1
-YUVJPEG=$2
-JPEGYUV=$3
-YUV2YUV4MPEG=$4
-DUMP_PSNR=$5
-DUMP_PSNRHVS=$6
-DUMP_SSIM=$7
-DUMP_FASTSSIM=$8
-FILE=$9
+if [ -z $RD_COLLECT_SUB ]; then
+  echo "Please use: $(dirname $0)/rd_collect.sh x264 *.y4m"
+  exit 1
+fi
+
+FILE=$1
 
 BASENAME=$(basename $FILE)
 rm $BASENAME.out 2> /dev/null || true
 echo $BASENAME
-tail -n+3 $FILE > $BASENAME-in.yuv
+
 WIDTH=$(head -1 $FILE | cut -d\  -f 2 | tr -d 'W')
 HEIGHT=$(head -1 $FILE | cut -d\  -f 3 | tr -d 'H')
+PIXELS=$(($WIDTH*$HEIGHT))
 
-for x in {0..100}; do
-  $YUVJPEG $x "$WIDTH"x$HEIGHT $BASENAME-in.yuv $BASENAME.jpeg
-  $JPEGYUV $BASENAME.jpeg $BASENAME.yuv
+RANGE=$(seq 1 51)
+QSTR="--preset placebo --crf=\$x"
+
+for x in $RANGE; do
+  $X264 --dump-yuv $BASENAME.yuv $(echo $QSTR | sed 's/\$x/'$x'/g') -o $BASENAME.x264 $FILE 2> $BASENAME-$x-enc.out > /dev/null
   $YUV2YUV4MPEG $BASENAME -w$WIDTH -h$HEIGHT -an0 -ad0 -c420mpeg2
-  PIXELS=$(($WIDTH*$HEIGHT))
-  SIZE=$(wc -c $BASENAME.jpeg | awk '{ print $1 }')
+  SIZE=$(wc -c $BASENAME.x264 | awk '{ print $1 }')
   PSNR=$($DUMP_PSNR $FILE $BASENAME.y4m 2> /dev/null | grep Total | tr -s ' ' | cut -d\  -f $((4+$PLANE*2)))
   PSNRHVS=$($DUMP_PSNRHVS $FILE $BASENAME.y4m 2> /dev/null | grep Total | tr -s ' ' | cut -d\  -f $((4+$PLANE*2)))
   SSIM=$($DUMP_SSIM $FILE $BASENAME.y4m 2> /dev/null | grep Total | tr -s ' ' | cut -d\  -f $((4+$PLANE*2)))
   FASTSSIM=$($DUMP_FASTSSIM -c $FILE $BASENAME.y4m 2> /dev/null | grep Total | tr -s ' ' | cut -d\  -f $((4+$PLANE*2)))
-  rm $BASENAME.jpeg $BASENAME.yuv $BASENAME.y4m
+  rm $BASENAME.x264 $BASENAME.y4m $BASENAME.yuv $BASENAME-$x-enc.out
   echo $x $PIXELS $SIZE $PSNR $PSNRHVS $SSIM $FASTSSIM >> $BASENAME.out
   #tail -1 $BASENAME.out
 done
-
-rm $BASENAME-in.yuv
